+++ olive/systems/python_environment/python_environment_system.py
@@ -16,7 +16,7 @@ from typing import Any, Dict, List, Optional, Union
 import numpy as np
 import torch
 
-from olive.common.utils import run_subprocess
+from olive.common.utils import get_package_name_from_ep, run_subprocess
 from olive.evaluator.metric import (
     Metric,
     MetricResult,
@@ -31,7 +31,6 @@ from olive.passes.olive_pass import Pass
 from olive.systems.common import SystemType
 from olive.systems.olive_system import OliveSystem
 from olive.systems.system_config import PythonEnvironmentTargetUserConfig
-from olive.utils import get_package_name_from_ep
 
 logger = logging.getLogger(__name__)
